requirements.txt
streamlit
metaphor-python
openai
googlesearch-python
paperswithcode-client
requests
pandas